refactor: move tariff engine into package
