+++ tests/test_tariff_rules.py
@@ -6,7 +6,7 @@ import pytest
 
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
-from tariff_engine import (
+from bot_alista.tariff_engine import (
     calc_breakdown_rules,
     calc_breakdown_with_mode,
     calc_clearance_fee_rub,
